Move ResourceData classes to Celbridge.Entities ahead of refactor
