+++ Celbridge/Celbridge/Celbridge.csproj
@@ -1,6 +1,6 @@
 <Project Sdk="Uno.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>net8.0;net8.0-maccatalyst;net8.0-windows10.0.22621;net8.0-desktop</TargetFrameworks>
+    <TargetFrameworks>net8.0-windows10.0.22621;</TargetFrameworks>
     <UnoDisableVSWarnNetIsFirst>true</UnoDisableVSWarnNetIsFirst>
     
     <OutputType>Exe</OutputType>
