+++ Celbridge/Modules/Workspace/Celbridge.Entities/Celbridge.Entities.csproj
@@ -29,6 +29,7 @@
 
   <ItemGroup>
     <PackageReference Include="CommunityToolkit.Diagnostics" />
+    <PackageReference Include="NewtonSoft.Json" />
     <PackageReference Include="JsonPatch.Net" />
     <PackageReference Include="JsonPointer.Net" />
     <PackageReference Include="JsonSchema.Net" />
