--- Celbridge/Modules/Workspace/Celbridge.ResourceData/Celbridge.ResourceData.csproj
@@ -1,34 +1,0 @@
-<Project Sdk="Uno.Sdk">
-  <PropertyGroup>
-    <TargetFrameworks>net8.0;net8.0-maccatalyst;net8.0-windows10.0.22621;net8.0-desktop</TargetFrameworks>
-    <UnoDisableVSWarnNetIsFirst>true</UnoDisableVSWarnNetIsFirst>
-    <UnoSingleProject>true</UnoSingleProject>
-    <OutputType>Library</OutputType>
-    <GenerateLibraryLayout>true</GenerateLibraryLayout>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-
-    <UnoFeatures>
-      CSharpMarkup;
-      Lottie;
-      Hosting;
-      Toolkit;
-      Logging;
-      Mvvm;
-      Configuration;
-      Localization;
-      ThemeService;
-    </UnoFeatures>
-
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="CommunityToolkit.Diagnostics" />
-    <PackageReference Include="Newtonsoft.Json" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\CoreServices\Celbridge.BaseLibrary\Celbridge.BaseLibrary.csproj" />
-  </ItemGroup>
-
-</Project>
