--- Celbridge/Modules/Workspace/Celbridge.ResourceData/GlobalUsings.cs
@@ -1,2 +1,0 @@
-global using Celbridge.Foundation;
-
