+++ Celbridge/Modules/Workspace/Celbridge.Workspace/Celbridge.Workspace.csproj
@@ -36,7 +36,6 @@
     <ProjectReference Include="..\Celbridge.Explorer\Celbridge.Explorer.csproj" />
     <ProjectReference Include="..\Celbridge.GenerativeAI\Celbridge.GenerativeAI.csproj" />
     <ProjectReference Include="..\Celbridge.Inspector\Celbridge.Inspector.csproj" />
-    <ProjectReference Include="..\Celbridge.ResourceData\Celbridge.ResourceData.csproj" />
     <ProjectReference Include="..\Celbridge.Scripting\Celbridge.Scripting.csproj" />
     <ProjectReference Include="..\Celbridge.Status\Celbridge.Status.csproj" />
   </ItemGroup>
