+++ Celbridge/Modules/Workspace/Celbridge.Workspace/Module.cs
@@ -20,7 +20,6 @@ public class Extension : IModule
         Documents.ServiceConfiguration.ConfigureServices(config);
         Explorer.ServiceConfiguration.ConfigureServices(config);
         Inspector.ServiceConfiguration.ConfigureServices(config);
-        ResourceData.ServiceConfiguration.ConfigureServices(config);
         Entities.ServiceConfiguration.ConfigureServices(config);
         Scripting.ServiceConfiguration.ConfigureServices(config);
         Status.ServiceConfiguration.ConfigureServices(config);
